finished authorization

+++ app.py
@@ -3,6 +3,7 @@ from db import db, Item, User, Location
 from flask import Flask, request
 import json
 import os
+import datetime
 
 app = Flask(__name__)
 db_filename = "lostandfound.db"
@@ -15,14 +16,55 @@ db.init_app(app)
 with app.app_context():
     db.create_all()
 
-
-#routes here
-
+# helper methods
 def success_response(data, code = 200):
     return json.dumps(data), code
 
 def failure_response(message, code = 400):
     return json.dumps({"error": message}), code
+
+def extract_token(request):
+    """
+    Helper function that extracts the token from the header of a request
+    """
+    auth_header = request.headers.get("Authorization")
+    if auth_header is None:
+        return False, failure_response("Missing authorization header.", 400)
+    
+    # Header looks like "Authorization: Bearer <token>"
+    bearer_token = auth_header.replace("Bearer ", "").strip()
+
+    if bearer_token is None or not bearer_token:
+        return False, failure_response("Invalid authorization header", 400)
+
+    return True, bearer_token
+
+def verify_user(email, password):
+    """
+    helper function that checks user provided email and password are valid
+    """
+    user = User.query.filter(User.email == email).first()
+
+    if user is None:
+        return False, None
+
+    return user.verify_password(password), user
+
+def verify_session_token(request):
+    """
+    helper function that verifies if a session token is valid and grants access to protected endpoints
+    """
+    success, session_token = extract_token(request)
+
+    if not success:
+        return False, None
+
+    user = User.query.filter(User.session_token == session_token).first()
+
+    if user is None:
+        return False, None
+    
+    return True, user
 
 # User routes
 
@@ -42,19 +84,24 @@ def create_user():
     body = json.loads(request.data)
     name = body.get("name")
     email = body.get("email")
-    # password = body.get("password")
-    # verify_password = body.get("verify_password")
-    # or not password or not verify_password:
-    if name == None or email == None:
+    password = body.get("password")
+    
+    if name == None or email == None or password == None:
         return failure_response("Missing required fields", 400)
-    # if password != verify_password:
-    #     return failure_response("Passwords do not match", 400)
+    
     if User.query.filter_by(email = email).first():
         return failure_response("Email already in use", 400)
-    user = User(name = name, email = email)
+    
+    user = User(name = name, email = email, password = password)
+
     db.session.add(user)
     db.session.commit()
-    return success_response(user.serialize(), 201)
+
+    return success_response({
+        "session_token": user.session_token,
+        "session_expiration": str(user.session_expiration),
+        "update_token": user.update_token
+    })
     
 
 @app.route("/api/users/<int:user_id>/", methods=["GET"])
@@ -68,25 +115,92 @@ def get_user(user_id):
     return success_response(user.serialize())
 
 
-@app.route("/api/users/<int:user_id>/", methods=["DELETE"])
-def delete_user(user_id):
+@app.route("/api/users/delete/", methods=["DELETE"])
+def delete_user():
     """
-    deletes a user 
+    deletes a user
     """
-    user = User.query.filter_by(id = user_id).first()
-    if not user:
-        return failure_response("User not found", 404)
+    success, user = verify_session_token(request)    
+
+    if not success:
+        return failure_response("Could not verify session token!", 404)
+    
     db.session.delete(user)
     db.session.commit()
     return success_response(user.serialize())
 
-# auth route here
 @app.route("/api/users/login/", methods = ["POST"])
 def login():
     """
-    attempts to verify user and returns session token if successful
+    attempts to verify user and returns credentials if successful
     """
-    pass
+    body = json.loads(request.data)
+    email = body.get("email")
+    password = body.get("password")
+    
+    if email is None or password is None:
+        return failure_response("Missing email or password", 400)
+
+    success, user = verify_user(email, password)
+
+    if not success:
+        return failure_response("Incorrect email or password", 401)
+
+    user.renew_session()
+    db.session.commit()
+
+    return success_response(
+        {"session_token": user.session_token,
+        "session_expiration": str(user.session_expiration),
+        "update_token": user.update_token}
+    )
+
+@app.route("/api/users/session/", methods = ["POST"])
+def update_session():
+    """
+    updates a users session when it expires
+    """
+    success, update_token = extract_token(request)
+
+    if not success:
+        return failure_response("Could not extract update token", 400)
+
+    optional_user = User.query.filter(User.update_token == update_token).first()
+
+    if optional_user is None:
+        return failure_response("Invalid update token!", 400)
+
+    optional_user.renew_session()
+    db.session.commit()
+
+    return success_response({
+        "session_token": optional_user.session_token,
+        "session_expiration": str(optional_user.session_expiration),
+        "update_token": optional_user.update_token
+    })
+
+@app.route("/api/users/logout/", methods = ["POST"])
+def logout():
+    """
+    logs out a user
+    """
+
+    success, session_token = extract_token(request)
+
+    if not success:
+        return failure_response("Could not extract session token")
+    
+    user = User.query.filter(User.session_token == session_token).first()
+
+    if user is None or not user.verify_session_token(session_token):
+        return failure_response("Invalid session token!")
+
+    user.session_token = ""
+    user.session_expiration = datetime.datetime.now()
+    user.update_token = ""
+    db.session.commit()
+
+    return success_response({"message": "You have successfully logged out"})
 
 # Item routes
 
@@ -140,11 +254,21 @@ def get_item(item_id):
 @app.route("/api/items/<int:item_id>/", methods=["DELETE"])
 def delete_item(item_id):
     """
-    deletes an item based on item id
+    deletes an item based on item id (protected endpoint)
     """
     item = Item.query.filter_by(id = item_id).first()
     if item is None:
         return failure_response("Item does not exist!", 404)
+
+    success, user = verify_session_token(request)    
+
+    if not success:
+        return failure_response("Could not verify session token!", 404)
+
+    if user.id != item.user_id:
+        return failure_response("Could not verify item belongs to user", 404)
+
+    
     
     db.session.delete(item)
     db.session.commit()
@@ -154,11 +278,19 @@ def delete_item(item_id):
 @app.route("/api/items/<int:item_id>/", methods = ["POST"])
 def update_item(item_id):
     """
-    updates an item given item id; cannot edit user or location
+    updates an item given item id; cannot edit user or location (protected endpoint)
     """
     item = Item.query.filter_by(id = item_id).first()
     if item is None:
         return failure_response("Item does not exist!", 404)
+
+    success, user = verify_session_token(request)    
+
+    if not success:
+        return failure_response("Could not verify session token!", 404)
+
+    if user.id != item.user_id:
+        return failure_response("Could not verify item belongs to user", 404)
 
     body = json.loads(request.data)
     description = body.get("description")
